Move AI Guard test helpers into llm/utils.py
tests/test_the_test/test_conventions.py::test_utils only allows test_*.py,
utils.py, conftest.py and __init__.py under tests/, so the new
ai_guard_utils.py failed TEST_THE_TEST. Fold the helpers into the existing
tests/integration_frameworks/llm/utils.py instead.

JIRA: APPSEC-69629
